+++ packages/ui/package.json
@@ -6,6 +6,7 @@
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
+    "./filter": "./src/components/filter.tsx",
     "./index.css": "./src/index.css",
     "./progress": "./src/components/progress.tsx",
     "./tooltip": "./src/components/tooltip.tsx"
